Move duration opts into an opts package
They have nothing to do with service and could be used on their own.
